requirements.txt
streamlit==1.32.0
torch
torchvision
opencv-python
matplotlib
seaborn
scikit-learn
tqdm
Pillow
numpy